Move more implementation code out of `dora-message` into `dora-core`
Ensures that we need to update `dora-message` less often.
