+++ libraries/message/src/lib.rs
@@ -22,6 +22,9 @@ pub mod node_to_daemon;
 pub mod cli_to_coordinator;
 pub mod coordinator_to_cli;
 
+pub use arrow_data;
+pub use arrow_schema;
+
 pub type DataflowId = uuid::Uuid;
 
 fn current_crate_version() -> semver::Version {
